+++ test/uncomplicate/neanderthal/opencl_test.clj
@@ -2,14 +2,14 @@
   (:require [midje.sweet :refer :all]
             [uncomplicate.commons.core :refer [with-release]]
             [uncomplicate.neanderthal
-             [protocols :refer [factory data-accessor]]
+             [protocols :refer [native-factory data-accessor]]
              [core :refer :all]
              [opencl :refer :all]
              [math :refer [pow]]])
   (:import [uncomplicate.neanderthal.protocols Block]))
 
 (defn test-clblock [ocl-factory]
-  (let [host-factory (factory (data-accessor ocl-factory))
+  (let [host-factory (native-factory (data-accessor ocl-factory))
         m 33
         n (long (+ 1000 (pow 2 12)))
         cnt n
@@ -57,7 +57,7 @@
 (defn test-blas1 [ocl-factory]
   (facts
    "BLAS methods"
-   (let [host-factory (factory (data-accessor ocl-factory))
+   (let [host-factory (native-factory (data-accessor ocl-factory))
          cnt (long (+ 1000 (pow 2 12)))
          x-magic 2
          y-magic 5]
@@ -86,7 +86,7 @@
 
        (transfer (axpy! 2 cl-x cl-y)) => (axpy! 2 host-x host-y)))
 
-   (let [host-factory (factory (data-accessor ocl-factory))
+   (let [host-factory (native-factory (data-accessor ocl-factory))
          cnt (long (+ 1000 (pow 2 12)))
          x-magic 2
          y-magic 5]
@@ -113,7 +113,7 @@
 (defn test-blas2 [ocl-factory]
   (facts
    "BLAS 2"
-   (let [host-factory (factory (data-accessor ocl-factory))
+   (let [host-factory (native-factory (data-accessor ocl-factory))
          m-cnt 2050
          n-cnt 337
          a-magic 3
@@ -138,7 +138,7 @@
 (defn test-blas3 [ocl-factory]
   (facts
    "BLAS 3"
-   (let [host-factory (factory (data-accessor ocl-factory))
+   (let [host-factory (native-factory (data-accessor ocl-factory))
          m-cnt 123
          k-cnt 456
          n-cnt 789]